pyproject.toml
[project]
name = "benchflow"
version = "0.1.0"
description = "Scenario-based polyglot database benchmark platform"
requires-python = ">=3.10"

dependencies = [
    "typer>=0.12",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "jinja2>=3.1",
    "plotly>=5.0",
    "numpy>=1.24",
    "psycopg[binary]>=3.1",
    "sqlalchemy>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
]

[project.scripts]
bench = "benchflow.cli.main:app"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["benchflow*"]
